+++ C/matrix_op.h
@@ -5,7 +5,7 @@ extern "C" {
 #endif
 #include "macros.h"
 typedef struct _matrix_t {
-  float* data;
+  double* data;
   int rows;
   int cols;
   int max_size;
@@ -25,8 +25,8 @@ int elem_wise_mult(matrix_t* a, matrix_t* b);
 int elem_wise_div(matrix_t* a, matrix_t* b, matrix_t* ret);
 int add_bias(matrix_t* a, matrix_t* b);
 int mult_bias(matrix_t* a, matrix_t* b);
-int add_scalar(matrix_t* a, float b);
-int mult_scalar(matrix_t* a, float b);
+int add_scalar(matrix_t* a, double b);
+int mult_scalar(matrix_t* a, double b);
 int neg(matrix_t* a);
 int inverse(matrix_t* a);
 int square_root(matrix_t* a, matrix_t* ret);
@@ -43,11 +43,11 @@ int matmul_mkl(matrix_t* a, matrix_t* b, matrix_t* ret);
 #endif
 matrix_t* transpose(matrix_t* a);
 
-float mean(matrix_t* a);
+double mean(matrix_t* a);
 
 int equal(matrix_t* a, matrix_t* b);
 
-int xavier_init(matrix_t* a, float gain);
+int xavier_init(matrix_t* a, double gain);
 int truncated_normal_init(matrix_t* a);
 int zero_init(matrix_t* a);
 int ones_init(matrix_t* a);
@@ -58,7 +58,7 @@ int free_matrix(matrix_t* t);
 int contains_nan(matrix_t* t);
 int print_matrix(matrix_t* t, int all);
 int copy_matrix(matrix_t* dst, matrix_t* src);
-int any_larger(matrix_t* t, float thres);
+int any_larger(matrix_t* t, double thres);
 int augment_space(matrix_t* t, int rows, int cols);
 int* shuffle_row_wise(matrix_t* t, int* idx);
 matrix_t* new_matrix(int rows, int cols);
@@ -67,7 +67,7 @@ matrix_t* slice_row_wise(matrix_t* t, int start, int end);
 matrix_t* slice_col_wise(matrix_t* t, int start, int end);
 matrix_t* concatenate(matrix_t* a, matrix_t* b, int axis);
 matrix_t* matrix_clone(matrix_t* a);
-int clip(matrix_t* a, float low, float high);
+int clip(matrix_t* a, double low, double high);
 
 
 #ifdef __cplusplus
