+++ C/model.h
@@ -26,12 +26,12 @@ int print_network(model* m);
 
 // initialize all the memories for cache according to batch size
 // run through all the sample and update model
-float fit(model* m, matrix_t* x, matrix_t* y, int batch_size, int epoch, float learning_rate, int shuffle, int auto_update);
+double fit(model* m, matrix_t* x, matrix_t* y, int batch_size, int epoch, double learning_rate, int shuffle, int auto_update);
 int predict(model* m, matrix_t* x);
-float eval(model* m, matrix_t* x, matrix_t* y, matrix_t* min_max);
+double eval(model* m, matrix_t* x, matrix_t* y, matrix_t* min_max);
 
 int model_backward(model* m, matrix_t* grad);
-int perform_update(model* m, float learning_rate);
+int perform_update(model* m, double learning_rate);
 int init_adam(model* m);
 
 int init_caches(model* m, int batch_size);
