+++ C/multi_agents/mpi_utils.h
@@ -6,7 +6,7 @@
 #include "mpi.h"
 
 int mpi_init();
-int mpi_perform_update(model* m, float lr, int scale);
+int mpi_perform_update(model* m, double lr, int scale);
 int mpi_update_normalizer(normalizer* norm, matrix_t** data, int count);
 int mpi_sync(model* m);
 int mpi_check_sync(model* m);
